# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/brothers/{brotherID}/statuses": {
            "patch": {
                "description": "Deletes the status of the specified brother for the specified semester.",
                "tags": [
                    "Statuses"
                ],
                "summary": "Deletes the status of brother for specified semester",
                "parameters": [
                    {
                        "type": "string",
                        "description": "brotherID",
                        "name": "brotherID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "semesterID",
                        "name": "semesterID",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "description": "body",
                        "name": "status",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    }
                }
            }
        },
        "/api/brothers/{brotherID}/statuses/{semesterID}": {
            "delete": {
                "description": "Deletes the status of the specified brother for the specified semester.",
                "tags": [
                    "Statuses"
                ],
                "summary": "Deletes the status of brother for specified semester",
                "parameters": [
                    {
                        "type": "string",
                        "description": "brotherID",
                        "name": "brotherID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "semesterID",
                        "name": "semesterID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    }
                }
            }
        },
...
        "/api/brothers/{id}/statuses": {
            "get": {
                "description": "Get all status recorded for Brother",
                "tags": [
                    "Brothers"
                ],
                "summary": "Get status history of a Brother",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Brother ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new status record for a Brother",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Brothers"
                ],
                "summary": "Create status record for Brother",
                "parameters": [
                    {
                        "description": "Values for new record",
                        "name": "body_params",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.CreateBrotherStatus.RequestBody"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "Brother ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    }
                }
            }
        },
...
        "/api/statuses": {
            "get": {
                "description": "Get all valid status labels (e.g.: \"Active\")",
                "tags": [
                    "Statuses"
                ],
                "summary": "Get status labels",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handlers.CreateBrotherStatus.RequestBody": {
            "type": "object",
            "properties": {
                "brotherID": {
                    "type": "integer"
                },
                "semesterID": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "models.APIResponse": {
            "description": "JSON response format for all API calls",
            "type": "object",
            "properties": {
                "data": {
                    "description": "Omit if nil"
                },
                "message": {
                    "description": "Omit if nil",
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
    }
}